% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-circular.R
\name{add_circular_tree}
\alias{add_circular_tree}
\title{Add a circular dendrogram to a circular plot}
\usage{
add_circular_tree(
  obj,
  which = c("col", "row"),
  tree = NULL,
  dist_method = "euclidean",
  hclust_method = "complete",
  dist_fn = NULL,
  color = "grey30",
  line_size = 0.3
)
}
\arguments{
\item{obj}{A \code{ggcomplex} object with \code{plot_type = "circular"}.}

\item{which}{\code{"col"} or \code{"row"}.}

\item{tree}{Optional pre-computed \code{hclust} / \code{dendrogram} object.}

\item{dist_method}{Distance method.}

\item{hclust_method}{Agglomeration method.}

\item{dist_fn}{Optional custom distance function.}

\item{color}{Branch colour.}

\item{line_size}{Branch line width.}
}
\value{
Updated \code{ggcomplex} object.
}
\description{
Add a circular dendrogram to a circular plot
}
