% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{apply_anno_styling}
\alias{apply_anno_styling}
\title{Apply palette, legend, extra layers, and clean theme}
\usage{
apply_anno_styling(
  p,
  palette,
  continuous_palette,
  show_legend,
  extra_layers,
  discrete_axis = "x",
  anno_title = NULL,
  panel_border_color = NA,
  panel_border_width = 0.5
)
}
\description{
Apply palette, legend, extra layers, and clean theme
}
\keyword{internal}
