% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-main.R
\name{apply_split_gaps}
\alias{apply_split_gaps}
\title{Insert visual gaps between row/col split groups}
\usage{
apply_split_gaps(obj)
}
\description{
Insert visual gaps between row/col split groups
}
\keyword{internal}
