% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{assemble_plot}
\alias{assemble_plot}
\title{Assemble the composite plot}
\usage{
assemble_plot(obj)
}
\arguments{
\item{obj}{A \code{ggcomplex} object.}
}
\value{
An \code{aplot} object ready for printing.
}
\description{
Reads the \code{plots} and \code{params} slots of a \code{ggcomplex} object and uses
\code{\link[aplot:plot-insertion]{aplot::insert_left()}}, \code{\link[aplot:plot-insertion]{aplot::insert_right()}}, \code{\link[aplot:plot-insertion]{aplot::insert_top()}}, and
\code{\link[aplot:plot-insertion]{aplot::insert_bottom()}} to compose the final figure.
}
