% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{build_geom_args}
\alias{build_geom_args}
\title{Build a named list of geom arguments from the convenience parameters}
\usage{
build_geom_args(
  fill,
  color,
  alpha,
  border_color,
  border_width,
  mapping = NULL,
  dot_args = list()
)
}
\description{
Build a named list of geom arguments from the convenience parameters
}
\keyword{internal}
