% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ggcomplex.R
\name{is_ggcomplex}
\alias{is_ggcomplex}
\title{Test if an object is a ggcomplex}
\usage{
is_ggcomplex(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical.
}
\description{
Test if an object is a ggcomplex
}
\keyword{internal}
