% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{relevel_tidy}
\alias{relevel_tidy}
\title{Re-level tidy data after clustering}
\usage{
relevel_tidy(tidy, new_row_order = NULL, new_col_order = NULL)
}
\arguments{
\item{tidy}{A tibble produced by \code{\link[=matrix_to_tidy]{matrix_to_tidy()}}.}

\item{new_row_order}{New row order (character), or \code{NULL} to keep current.}

\item{new_col_order}{New column order (character), or \code{NULL} to keep current.}
}
\value{
Updated tibble.
}
\description{
Updates the factor levels in \code{tidy} to match a new ordering vector.
}
\keyword{internal}
