% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{set_title}
\alias{set_title}
\title{Set global titles for a ggcomplex plot}
\usage{
set_title(obj, title = NULL, subtitle = NULL, caption = NULL)
}
\arguments{
\item{obj}{A \code{ggcomplex} object.}

\item{title}{Plot title.}

\item{subtitle}{Plot subtitle.}

\item{caption}{Plot caption (bottom-right).}
}
\value{
Updated \code{ggcomplex} object.
}
\description{
Adds a title, subtitle, and/or caption that will be rendered above/below
the entire assembled composite figure.
}
