% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset_cols}
\alias{subset_cols}
\title{Subset columns of a ggcomplex object}
\usage{
subset_cols(obj, cols)
}
\arguments{
\item{obj}{A \code{ggcomplex} object.}

\item{cols}{Character vector of column names to keep.}
}
\value{
Updated \code{ggcomplex} object.
}
\description{
Re-creates the main plot with only the specified columns.
}
