% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset_rows}
\alias{subset_rows}
\title{Subset rows of a ggcomplex object}
\usage{
subset_rows(obj, rows)
}
\arguments{
\item{obj}{A \code{ggcomplex} object.}

\item{rows}{Character vector of row names to keep.}
}
\value{
Updated \code{ggcomplex} object.
}
\description{
Re-creates the main plot with only the specified rows.
}
